src, sandbox2D: fix Renderer2D bugs

+++ sandbox2D/res/shaders/2D_vs.glsl
@@ -3,11 +3,11 @@
 layout (location = 0) in vec3 a_pos;
 layout (location = 1) in vec4 a_color;
 layout (location = 2) in vec2 a_tex_coord;
-layout (location = 3) in uint a_tex_index;
+layout (location = 3) in float a_tex_index;
 
 out vec4 color;
 out vec2 texture_coordinate;
-flat out uint texture_index;
+flat out float texture_index;
 
 uniform mat4 view;
 uniform mat4 projection;
